setup: moved to installer and refactored
This moves the setup.py script ot installer.py, adding cli options and
refactoring some of the functions to not need the 'pi' user.

This also improves on the error detection and better handles the
pip/apt installation (also through requires.txt for pip, and through
script for apt so it only installs things once).

This removed some duplicated dependencies too.